requirements.txt
aiofiles==23.2.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1717618050233/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.2
conda @ file:///croot/conda_1715635703388/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro @ file:///croot/distro_1714488253808/work
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.9.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.28.1
idna==2.10
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.10.0
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpickle==3.3.0
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1718132535249/work
mistune==3.1.1
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
openai==1.46.0
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prometheus_client==0.21.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.6.0
torchattacks==3.5.1
torchvision==0.21.0
tornado==6.4.2
tqdm @ file:///croot/tqdm_1716395931952/work
traitlets==5.14.3
transformers==4.44.2
triton==3.2.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.4
wheel==0.43.0
zstandard @ file:///croot/zstandard_1714677652653/work
